refactor: CouponService로 통일
