+++ src/main/java/coupon/coupon/application/CouponService.java
@@ -8,12 +8,20 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class CouponCommandService {
+public class CouponService {
 
     private final CouponRepository couponRepository;
 
     @Transactional
     public void create(Coupon coupon) {
         couponRepository.save(coupon);
+    }
+
+    @Transactional(readOnly = true)
+    public Coupon getCoupon(Long couponId) {
+        String message = "존재하지 않는 쿠폰입니다. id: %d".formatted(couponId);
+
+        return couponRepository.findById(couponId)
+                .orElseThrow(() -> new IllegalArgumentException(message));
     }
 }
